+++ HISTORY.rst
@@ -8,7 +8,9 @@ History
 * Daemon mode
 * Logging (+ log files)
 * Global config file (minipyp.conf)
-* Directory-specific config files (.minipyp)
+* Major rewrite of core
+* Better Python page handler
+* Better getter/setter for headers & query
 
 0.1.0b1
 -------